Refactor register pages to use a shared design and move components into pages
