--- application/account-management/WebApp/pages/login/_layout.tsx
@@ -1,9 +1,0 @@
-import { createFileRoute, Outlet } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/login/_layout")({
-  component: LoginLayout
-});
-
-export default function LoginLayout() {
-  return <Outlet />;
-}
