+++ application/account-management/WebApp/pages/register/-state/actions.ts
@@ -1,6 +1,7 @@
 import { i18n } from "@lingui/core";
 import { getApiError, getFieldErrors } from "@repo/infrastructure/api/ErrorList";
 import { accountManagementApi } from "@/shared/lib/api/client";
+import type { FetchResponse } from "openapi-fetch";
 
 const VALIDATION_LIFETIME = 1000 * 60 * 5; // 5 minutes
 
@@ -17,6 +18,7 @@ interface Registration {
 export const registration: Registration = { current: undefined };
 
 export interface State {
+  error: boolean;
   success?: boolean;
   message?: string | null;
   errors?: { [key: string]: string | string[] };
@@ -31,27 +33,41 @@ export async function startAccountRegistration(_: State, formData: FormData): Pr
   });
 
   if (!result.response.ok) {
-    const apiError = getApiError(result);
-    let errorState = { success: false, message: apiError.title, errors: getFieldErrors(apiError.Errors) };
+    const errorState = convertResponseErrorToErrorState(result);
     console.log(errorState);
     return errorState;
   }
 
   try {
-    const location = result.response.headers.get("Location")!;
-    const accountRegistrationId = location.split("/").pop()!;
+    const accountRegistrationId = extractRegistrationIdFromHeader(result.response.headers);
+
+    if (!accountRegistrationId) {
+      return {
+        error: true,
+        success: false,
+        message: i18n.t("An error occured when trying to start Account registration.")
+      };
+    }
 
     registration.current = { accountRegistrationId, email, expireAt: new Date(Date.now() + VALIDATION_LIFETIME) };
 
-    return { success: true };
+    return { error: false, success: true };
   } catch (e) {
-    return { success: false, message: i18n.t("An error occured when trying to start Account registration.") };
+    return {
+      error: true,
+      success: false,
+      message: i18n.t("An error occured when trying to start Account registration.")
+    };
   }
 }
 
 export async function completeAccountRegistration(_: State, formData: FormData): Promise<State> {
   const oneTimePassword = formData.get("oneTimePassword") as string;
-  const accountRegistrationId = registration.current?.accountRegistrationId!;
+  const accountRegistrationId = registration.current?.accountRegistrationId;
+
+  if (!accountRegistrationId) {
+    return { error: true, success: false, message: i18n.t("Account registration is not started.") };
+  }
 
   try {
     const result = await accountManagementApi.POST("/api/account-management/account-registrations/{id}/complete", {
@@ -60,14 +76,29 @@ export async function completeAccountRegistration(_: State, formData: FormData):
     });
 
     if (!result.response.ok) {
-      const apiError = getApiError(result);
-      let errorState = { success: false, message: apiError.title, errors: getFieldErrors(apiError.Errors) };
+      const errorState = convertResponseErrorToErrorState(result);
       console.log(errorState);
       return errorState;
     }
 
-    return { success: true };
+    return { error: false, success: true };
   } catch (e) {
-    return { success: false, message: i18n.t("An error occured when trying to complete Account registration.") };
+    return {
+      error: true,
+      success: false,
+      message: i18n.t("An error occured when trying to complete Account registration.")
+    };
   }
+}
+
+type MediaType = `${string}/${string}`;
+
+function convertResponseErrorToErrorState<T, O, M extends MediaType>(result: FetchResponse<T, O, M>): State {
+  const apiError = getApiError(result);
+  return { error: true, success: false, message: apiError.title, errors: getFieldErrors(apiError.Errors) };
+}
+
+function extractRegistrationIdFromHeader(headers: Headers): string | undefined {
+  const location = headers.get("Location");
+  return location?.split("/").pop();
 }
