+++ application/account-management/WebApp/shared/translations/locale/da-DK.po
@@ -18,28 +18,36 @@ msgid "yourname@example.com"
 msgstr "ditnavn@example.com"
 
 #. js-lingui-explicit-id
-msgid "Email"
-msgstr "Email"
-
-#. js-lingui-explicit-id
 msgid "An error occured when trying to start Account registration."
 msgstr "Der opstod en fejl. Kunne ikke starte kontooprettelse."
+
+#. js-lingui-explicit-id
+msgid "Account registration is not started."
+msgstr ""
 
 #. js-lingui-explicit-id
 msgid "An error occured when trying to complete Account registration."
 msgstr "Der opstod en fejl. Kunne ikke færdigøre kontooprettelse."
 
 #. js-lingui-explicit-id
-msgid "subdomain"
-msgstr "subdomæne"
+msgid "Email"
+msgstr "Email"
 
 #. js-lingui-explicit-id
 msgid "Subdomain"
 msgstr "Subdomain"
 
 #. js-lingui-explicit-id
+msgid "subdomain"
+msgstr "subdomæne"
+
+#. js-lingui-explicit-id
 msgid "Region"
 msgstr "Region"
+
+#. js-lingui-explicit-id
+msgid "This is the region where your data is stored"
+msgstr ""
 
 #. js-lingui-explicit-id
 msgid "Screenshots of the dashboard project showing mobile versions"
@@ -64,23 +72,23 @@ msgstr "Har du allerede en konto?"
 msgid "Can't find your code? Check your spam folder"
 msgstr "Kan du ikke finde din kode? Tjek din spam-mappe"
 
-msgid "Continue"
-msgstr "Fortsæt"
+#~ msgid "Continue"
+#~ msgstr "Fortsæt"
 
-msgid "Create your account"
-msgstr "Opret din konto"
+#~ msgid "Create your account"
+#~ msgstr "Opret din konto"
 
 msgid "Did't receive the code? Resend"
 msgstr "Modtog ikke koden? Send igen"
 
-msgid "Don't have an account?"
-msgstr "Har du ikke en konto?"
+#~ msgid "Don't have an account?"
+#~ msgstr "Har du ikke en konto?"
 
-msgid "Email"
-msgstr "E-mail"
+#~ msgid "Email"
+#~ msgstr "E-mail"
 
-msgid "Enter your email below to log in."
-msgstr "Indtast din e-mail nedenfor for at logge ind"
+#~ msgid "Enter your email below to log in."
+#~ msgstr "Indtast din e-mail nedenfor for at logge ind"
 
 msgid "Enter your verification code"
 msgstr "Indtast din verifikationskode"
@@ -91,8 +99,8 @@ msgstr "Filtre"
 msgid "Here’s your overview of what’s going on."
 msgstr "Her er en oversigt over, hvad der sker"
 
-msgid "Hi! Welcome back"
-msgstr "Hej! Velkommen tilbage"
+#~ msgid "Hi! Welcome back"
+#~ msgstr "Hej! Velkommen tilbage"
 
 msgid "Log in"
 msgstr "Log ind"
@@ -100,17 +108,17 @@ msgstr "Log ind"
 msgid "Please check your email for a verification code sent to <0>{email}</0>"
 msgstr "Tjek venligst din e-mail for en verifikationskode sendt til <0>{email}</0>"
 
-msgid "Region"
-msgstr "Region"
+#~ msgid "Region"
+#~ msgstr "Region"
 
-msgid "Sign up"
-msgstr "Opret"
+#~ msgid "Sign up"
+#~ msgstr "Opret"
 
-msgid "Subdomain"
-msgstr "Subdomæne"
+#~ msgid "Subdomain"
+#~ msgstr "Subdomæne"
 
-msgid "This is the region where your data is stored"
-msgstr "Dine data vil blive gemt i denne region."
+#~ msgid "This is the region where your data is stored"
+#~ msgstr "Dine data vil blive gemt i denne region."
 
 msgid "Total Users"
 msgstr "Totalt antal brugere"
